requirements.txt
Django==4.1.3
djangorestframework
absl-py==1.2.0
altair==4.2.0
astunparse==1.6.3
attrs==22.1.0
blinker==1.5
cachetools==5.2.0
certifi==2022.6.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
decorator==5.1.1
entrypoints==0.4
flatbuffers==2.0.7
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.48.1
h5py==3.7.0
idna==3.3
importlib-metadata==4.12.0
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.16.0
keras==2.10.0
Keras-Preprocessing==1.1.2
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
nltk==3.7
numpy==1.23.3
oauthlib==3.2.1
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.4
pickle-mixin==1.0.2
Pillow==9.2.0
protobuf==3.19.5
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydeck==0.8.0b3
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
regex==2022.9.13
requests==2.28.1
requests-oauthlib==1.3.1
rich==12.5.1
rsa==4.9
scikit-learn==1.1.2
scipy==1.9.1
semver==2.13.0
six==1.16.0
smmap==5.0.0
streamlit==1.12.2
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-cpu==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
typing-extensions==4.3.0
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.12
validators==0.20.0
watchdog==2.1.9
Werkzeug==2.2.2
wrapt==1.14.1
yake
zipp==3.8.1
snscrape
seaborn
nltk
wordcloud
